requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0

# AI/NLP
google-generativeai==0.3.1
spacy==3.8.0

# Environment & Utils
python-dotenv==1.0.0

# Optional but recommended
# en_core_web_sm - spaCy model (install separately)
# python -m spacy download en_core_web_sm